% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{process_sensors_r_version}
\alias{process_sensors_r_version}
\title{Sensor Data R}
\usage{
process_sensors_r_version(person_location)
}
\arguments{
\item{person_location}{the csv file with sensor temperature data}
}
\value{
Graphs and tables with monthly, dayly, hourly and minute by minute average temperatures and heat index
}
\description{
This function allows you to input your temperature data from the sensor and analyze it for monthly, dayly, hourly and minute by minute average temperatures and heat index in addition to creating graphs. The output is an R script with tables and graphs.
}
